requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
requests>=2.28.2
pydantic>=1.10.7
python-gitlab>=3.13.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.38.0
google-genai>=0.1.0
faiss-cpu>=1.7.4
pygments>=2.15.0
aiohttp>=3.8.4
jinja2>=3.1.2 